Create StreamWish.ts
The streamwish is using fake png headers on their ts segments so we need mediaflow for this extractor. I already pr my finding to the mediaflow author and he is now merging both the fake steip bytes and also streamwish support for this extractor.

p.s this eztractor uses dynamic domin redirections so better leave the logic of domains intact if you dont want to spend your time lol.